Add endpoint that generates room
In preparation for generating the world, I build the first steps of the endpoint that adds rooms to the database.  The endpoint looks for size on the request body and is an endpoint that is only accessible to an administrator.  Next, I'll add the code that will connect the rooms together to form a dungeon.